# openapi.yaml
openapi: "3.0.3"
info:
  title: E-commerce POC API
  version: "1.0.0"
  description: |
    A Proof of Concept for a scalable e-commerce backend built with Python 3.12, FastAPI,
    and Pydantic v2.

    All data is held **in-memory** (no persistent database); storage resets on every restart.
    The layered architecture (API → Service → Repository → Domain Model) is designed for
    easy migration to PostgreSQL + SQLAlchemy.

    ## Authentication
    No authentication or authorisation is currently implemented. All endpoints are public.

    ## Error format
    Most errors return a standard FastAPI `{"detail": "..."}` body.
    Pydantic validation failures return `{"detail": [{loc, msg, type}]}`.
    See the `ErrorDetail` and `ValidationError` schemas in Components.
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/v1/products/:
    get:
      summary: Get all products
      description: Returns every product in the catalogue, each with its nested category object.
      operationId: read_products
      tags:
        - Products
      responses:
        "200":
          description: A list of products (may be empty).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
              examples:
                single_product:
                  summary: Single electronics product
                  value:
                    - id: 1
                      name: Smartphone
                      description: Latest model
                      price: 699.99
                      sku: PHN-123
                      category:
                        id: 1
                        name: Electronics
                        description: Gadgets and devices

    post:
      summary: Create a new product
      description: |
        Creates a new product and links it to an existing category.
        Returns **422** if the specified `category_id` does not exist.
      operationId: create_product
      tags:
        - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
            example:
              name: Smartphone
              description: Latest model
              price: 699.99
              sku: PHN-123
              category_id: 1
      responses:
        "201":
          description: Product created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/products/{product_id}:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    get:
      summary: Get a product by ID
      description: Retrieves a single product, including its nested category object.
      operationId: read_product
      tags:
        - Products
      responses:
        "200":
          description: Product found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update a product
      description: |
        Updates one or more fields of an existing product. All body fields are optional.
        If `category_id` is supplied, it must reference a valid category — otherwise **422**.
      operationId: update_product
      tags:
        - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
            example:
              name: Flagship Smartphone
              price: 799.99
      responses:
        "200":
          description: Product updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

    delete:
      summary: Delete a product
      description: Permanently removes a product from the catalogue.
      operationId: delete_product
      tags:
        - Products
      responses:
        "204":
          description: Product deleted. No response body.
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Customers ─────────────────────────────────────────────────────────────
...
components:
...
  parameters:
...
    ProductId:
      name: product_id
      in: path
      required: true
      description: Integer identifier of the product.
      schema:
        type: integer
        example: 1
...
  responses:
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
          example:
            detail: Category with id 99 not found
    UnprocessableEntity:
      description: |
        Validation error. Can be produced by:
        - **Pydantic** field validation (e.g. `price ≤ 0`, malformed email address).
        - **Service-layer** business-rule violations (e.g. unknown `category_id` on product create).

        Pydantic errors return the `ValidationError` shape; service-layer errors return the
        simpler `ErrorDetail` shape.
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/ValidationError"
              - $ref: "#/components/schemas/ErrorDetail"
          examples:
            pydantic_field_error:
              summary: Pydantic field validation failure
              value:
                detail:
                  - loc: ["body", "price"]
                    msg: Input should be greater than 0
                    type: greater_than
            business_rule_error:
              summary: Service-layer business rule violation
              value:
                detail: Category with id 99 not found

  # ── Schemas ────────────────────────────────────────────────────────────────
  schemas:
...
    Category:
      type: object
      description: A product category as stored in the system.
      required:
        - id
        - name
      properties:
        id:
          type: integer
          description: Auto-assigned unique identifier.
          example: 1
        name:
          type: string
          example: Electronics
        description:
          type: string
          nullable: true
          example: Gadgets and devices
...
    ProductCreate:
      type: object
      description: Payload for creating a new product.
      required:
        - name
        - price
        - sku
        - category_id
      properties:
        name:
          type: string
          description: Product name.
          example: Smartphone
        description:
          type: string
          nullable: true
          description: Optional product description.
          example: Latest model
        price:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
          description: Unit price. Must be strictly greater than zero.
          example: 699.99
        sku:
          type: string
          description: Stock Keeping Unit — unique product code.
          example: PHN-123
        category_id:
          type: integer
          description: ID of an existing category this product belongs to.
          example: 1
    ProductUpdate:
      type: object
      description: Payload for a partial product update. All fields are optional.
      properties:
        name:
          type: string
          nullable: true
          example: Flagship Smartphone
        description:
          type: string
          nullable: true
          example: Next-generation flagship
        price:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
          nullable: true
          description: Updated price. Must be strictly greater than zero if provided.
          example: 799.99
        sku:
          type: string
          nullable: true
          example: PHN-456
        category_id:
          type: integer
          nullable: true
          description: Updated category ID. Must reference an existing category if provided.
          example: 2
    Product:
      type: object
      description: A product as stored in the system, including its fully resolved category.
      required:
        - id
        - name
        - price
        - sku
        - category
      properties:
        id:
          type: integer
          description: Auto-assigned unique identifier.
          example: 1
        name:
          type: string
          example: Smartphone
        description:
          type: string
          nullable: true
          example: Latest model
        price:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
          example: 699.99
        sku:
          type: string
          example: PHN-123
        category:
          $ref: "#/components/schemas/Category"

    # ── Customer ──────────────────────────────────────────────────────────────
...
    ErrorDetail:
      type: object
      description: |
        Standard FastAPI `HTTPException` response body.
        Returned for 400, 404, and service-layer 422 errors.
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable description of the error.
          example: Category with id 99 not found
    ValidationErrorItem:
      type: object
      description: A single field-level error produced by Pydantic v2.
      properties:
        loc:
          type: array
          items:
            type: string
          description: JSON path segments identifying the invalid field.
          example: ["body", "price"]
        msg:
          type: string
          description: Human-readable error description.
          example: Input should be greater than 0
        type:
          type: string
          description: Pydantic internal error code.
          example: greater_than
    ValidationError:
      type: object
      description: Pydantic / FastAPI 422 response body containing one or more field-level errors.
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            $ref: "#/components/schemas/ValidationErrorItem"
          example:
            - loc: ["body", "price"]
              msg: Input should be greater than 0
              type: greater_than